backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.5
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.1
easyocr==1.7.1
motor==3.6.0

